Extract admission persistence into admission/persistence.rs
